+++ examples/AVR2ServoNIO/AVR2ServoNIO.ino
@@ -222,7 +222,7 @@ void userInitAll()
 }
 
 // ===== Process Consumer-eventIDs =====
-void pceCallback(unsigned int index) {
+void pceCallback(uint16_t index) {
 // Invoked when an event is consumed; drive pins as needed
 // from index of all events.
 // Sample code uses inverse of low bit of pattern to drive pin all on or all off.
@@ -317,7 +317,7 @@ void userHardReset() {}
 // Callback from a Configuration write
 // Use this to detect changes in the ndde's configuration
 // This may be useful to take immediate action on a change.
-void userConfigWritten(unsigned int address, unsigned int length, unsigned int func)
+void userConfigWritten(uint32_t address, uint16_t length, uint16_t func)
 {
   dPS("\nuserConfigWritten: Addr: ", (uint32_t)address);
   dPS(" Len: ", (uint16_t)length);
